compositor: expose tw_xdg structure.
Later we would use this it for binding xwayland_ready_listener.
