+++ compositor/config/config.c
@@ -46,7 +46,6 @@
 #include <taiwins/output_device.h>
 #include <taiwins/shell.h>
 
-#include "xdg.h"
 #include "bindings.h"
 #include "config.h"
