+++ compositor/desktop/layout_floating.c
@@ -29,6 +29,7 @@
 
 #include "layout.h"
 #include "workspace.h"
+#include "xdg.h"
 
 static void
 emplace_float(const enum tw_xdg_layout_command command,
