+++ compositor/desktop/layout_maximized.c
@@ -28,6 +28,7 @@
 
 #include "layout.h"
 #include "workspace.h"
+#include "xdg.h"
 
 static void
 emplace_maximized(const enum tw_xdg_layout_command command,
